+++ inst/include/simmer/process/arrival.h
@@ -1,5 +1,5 @@
 // Copyright (C) 2015-2016 Bart Smeets and Iñaki Ucar
-// Copyright (C) 2016-2018 Iñaki Ucar
+// Copyright (C) 2016-2019 Iñaki Ucar
 //
 // This file is part of simmer.
 //
@@ -276,7 +276,7 @@ namespace simmer {
 
     void unset_busy(double now) {
       set_remaining(status.busy_until - now);
-      set_busy(0);
+      set_busy(now);
     }
   };